requirements.txt
# Core Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# LLM & Embeddings
openai>=1.10.0
tiktoken>=0.5.0
sentence-transformers>=2.3.0
transformers>=4.37.0

# Orchestration
langgraph>=0.0.20
langchain>=0.1.0
langchain-openai>=0.0.5

# Data Processing
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0

# Vector Store & Search
faiss-cpu>=1.8.0
rank-bm25>=0.2.2

# Image Processing
Pillow>=10.0.0
pytesseract>=0.3.10

# Storage
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0
pyyaml>=6.0.0
aiofiles>=23.0.0

# Logging & Monitoring
loguru>=0.7.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
prometheus-client>=0.19.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.0.0
ruff>=0.1.0

# UI (Optional)
streamlit>=1.31.0

# web search
ddgs